NXDRIVE-1037: Ignore children of folder in duplicate error
Also, removed obsolete de-duplication tests and merged
all other revelant tests in one file.
